fix chooseoptions's state not updating after successfull vote
